+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.15] - 2026-03-26
+
+### Other
+- Update examples/planfile/calculator.py
+- Update examples/planfile/duplicate_test.py
+- Update examples/planfile/simple_test.py
+- Update examples/planfile/test_planfile_working.py
+- Update examples/planfile/test_report.py
+
 ## [0.1.14] - 2026-03-26
 
 ### Other